refactor: cleanup repository - rename and move sources - refactor code a bit. - simplify error handling - reformat all using prettier - update gitignore to latest github upstream - upgrade all deps to latest stable. - change scripts to match other repositories we have. - protect against not having GITHUB_TOKEN setup locally during tests. - refactor github actions configurations. ensure it runs the tests on win/mac as well. use actions/setup-node@v2.
