+++ .prettierignore
@@ -1,3 +1,3 @@
-dist/
-lib/
-node_modules/
+/dist
+/lib
+/node_modules
