+++ .prettierrc
@@ -1,12 +1,13 @@
 {
-    "printWidth": 120,
-    "singleQuote": true,
-    "overrides": [
-        {
-            "files": ["*.js", "*.ts", "*.tsx"],
-            "options": {
-                "tabWidth": 4
-            }
-        }
-    ]
+  "printWidth": 120,
+  "singleQuote": true,
+  "endOfLine": "auto",
+  "overrides": [
+    {
+      "files": ["*.js", "*.ts", "*.tsx"],
+      "options": {
+        "tabWidth": 4
+      }
+    }
+  ]
 }
