--- .vscode/settings.json
@@ -1,10 +1,0 @@
-{
-    "cSpell.words": [
-        "Octokit",
-        "mrkdwn",
-        "repos",
-        "subteam",
-        "thenable",
-        "vercel"
-    ]
-}
