+++ README.md
@@ -1,11 +1,9 @@
-<p align="center">
-  <a href="https://github.com/wixplosives/action-slack-message/actions"><img alt="typescript-action status" src="https://github.com/wixplosives/action-slack-message/workflows/build-test/badge.svg"></a>
-</p>
+[![Build Status](https://github.com/wixplosives/action-slack-message/workflows/tests/badge.svg)](https://github.com/wixplosives/action-slack-message/actions)
 
 # Slack Message Action
 
--   Use this action to send slack notification from github actions.
--   This action also let you extract the inner jobs link for each workflow
+- Use this action to send slack notification from github actions.
+- This action also let you extract the inner jobs link for each workflow
 
 This action includes compilation support, tests, a validation workflow, publishing, and versioning guidance.
 
@@ -32,22 +30,26 @@ $ npm test
 ```
 
 ## Run it to send slack message
+
 This is most basic example of how to run the action.
 
 ```
-- name: Notify slack 
-  uses: 'wixplosives/action-slack-message@master' 
+- name: Notify slack
+  uses: 'wixplosives/action-slack-message@master'
   with:
-      text: 'My text is so awesome'      
+      text: 'My text is so awesome'
       channel: 'testing-yarden-public'
       slack_token: ${{ secrets.SLACK_BOT_TOKEN }}
 ```
+
 ### Run it to notify about job status
-```if always()``` is required because we want to send status update even if action fails.
+
+`if always()` is required because we want to send status update even if action fails.
+
 ```
 - name: Notify slack
   if: always()
-  uses: 'wixplosives/action-slack-message@master'  
+  uses: 'wixplosives/action-slack-message@master'
   with:
       text: 'My text is so awesome'
       status: ${{job.status}}
@@ -55,15 +57,16 @@ This is most basic example of how to run the action.
       slack_token: ${{ secrets.SLACK_BOT_TOKEN }}
 ### **Custom job name**
 ```
+
 ### When you define a custom job name, like this
 
 ```yaml
 notify-custom-name:
-    runs-on: ${{ matrix.os }}
-    name: some custom name # <<< CUSTOM JOB NAME DEFINED HERE
-    strategy:
-        matrix:
-            os: [ubuntu-latest, windows-latest]
+  runs-on: ${{ matrix.os }}
+  name: some custom name # <<< CUSTOM JOB NAME DEFINED HERE
+  strategy:
+    matrix:
+      os: [ubuntu-latest, windows-latest]
 ```
 
 You need to pass as an argument the custom job name, as the following:
@@ -86,5 +89,5 @@ Add github.token as an environment variable
 - name: Notify slack
   uses: 'wixplosives/action-slack-message@master'
   env:
-      GITHUB_TOKEN: ${{ github.token }}
+    GITHUB_TOKEN: ${{ github.token }}
 ```
