+++ package.json
@@ -2,32 +2,34 @@
   "name": "action-slack-message",
   "version": "0.0.1",
   "private": true,
-  "description": "TypeScript template action",
+  "description": "Post messages to Slack from GitHub Actions",
   "main": "lib/main.js",
   "scripts": {
-    "build": "tsc -p tsconfig.build.json",
-    "format": "prettier --write **/*.ts",
-    "format-check": "prettier --check **/*.ts",
-    "lint": "eslint src/**/*.ts",
-    "pack": "ncc build --source-map --license licenses.txt",
-    "prebuild": "rimraf \"lib/*\"",
-    "prepack": "rimraf \"dist/*\"",
-    "test": "mocha \"tests/**/*.test.ts\"",
-    "all": "npm run build && npm run format && npm run lint && npm run pack && npm test"
+    "clean": "rimraf lib dist",
+    "prebuild": "npm run clean",
+    "build": "npm run build:ts && npm run bundle",
+    "build:ts": "tsc -p tsconfig.build.json",
+    "bundle": "ncc build --source-map --license licenses.txt",
+    "lint": "eslint . -f codeframe",
+    "typecheck": "tsc --noEmit",
+    "pretest": "npm run typecheck && npm run lint",
+    "test": "mocha \"test/**/*.test.ts\"",
+    "prettify": "npx prettier . --write"
   },
   "repository": {
     "type": "git",
-    "url": "git+https://github.com/actions/typescript-action.git"
+    "url": "git+https://github.com/wixplosives/action-slack-message.git"
   },
   "keywords": [
     "actions",
     "node",
-    "setup"
+    "slack",
+    "message"
   ],
   "author": "",
   "license": "MIT",
   "dependencies": {
-    "@actions/core": "^1.2.6",
+    "@actions/core": "^1.2.7",
     "@actions/github": "^4.0.0",
     "@octokit/core": "^3.4.0",
     "@slack/web-api": "^6.1.0"
@@ -36,17 +38,16 @@
     "@ts-tools/node": "^2.1.1",
     "@types/chai": "^4.2.16",
     "@types/mocha": "^8.2.2",
-    "@types/node": "^14.14.37",
-    "@typescript-eslint/eslint-plugin": "^4.21.0",
-    "@typescript-eslint/parser": "^4.21.0",
+    "@types/node": "14",
+    "@typescript-eslint/eslint-plugin": "^4.22.0",
+    "@typescript-eslint/parser": "^4.22.0",
     "@vercel/ncc": "^0.28.3",
     "chai": "^4.3.4",
     "eslint": "^7.24.0",
-    "eslint-config-prettier": "^8.1.0",
+    "eslint-config-prettier": "^8.2.0",
     "eslint-plugin-github": "^4.1.3",
     "eslint-plugin-no-only-tests": "^2.5.0",
     "mocha": "^8.3.2",
-    "prettier": "2.2.1",
     "rimraf": "^3.0.2",
     "typescript": "^4.2.4"
   }
