--- src/config/prod.ts
@@ -1,3 +1,0 @@
-module.exports = {
-    githubToken: process.env['GITHUB_TOKEN'],
-};
